Refactored utility functions to their own file and added documentation for App and StopWatch
