+++ src/StopWatch.tsx
@@ -1,49 +1,46 @@
 import React from 'react';
 import { Lap } from './App';
 import './StopWatch.css';
+import { formatTime } from './util';
 
-interface StopWatchProps {
-    laps: Lap[],
-    isRunning: boolean,
-    time: number,
-}
-
-export const breakdownTime = (time: number) => {
-    const ms = time % 1000;
-    const seconds = Math.floor(time / 1000) % 60;
-    const minutes = Math.floor(time / 1000 / 60) % 60;
-    const hours = Math.floor(time / 1000 / 60 / 60);
-
-    return {
-        ms,
-        seconds,
-        minutes,
-        hours,
-    }
-}
-
-export const formatTime = (time: number) => {
-    const { ms, seconds, minutes, hours } = breakdownTime(time);
-
-    const msString = ms.toString().padStart(3, '0');
-    const secondsString = seconds.toString().padStart(2, '0');
-    const minutesString = minutes.toString().padStart(2, '0');
-    const hoursString = hours.toString().padStart(2, '0');
-
-    return `${hoursString}:${minutesString}:${secondsString}:${msString}`;
-}
-
+/**
+ * Props for the LapList component
+ * @param laps the list of laps to display
+ * @param time the current time in milliseconds
+ */
 export interface LapListProps {
     laps: Lap[],
     time: number,
 }
 
+/**
+ * Props for the LapDisplay component
+ * @param faster true if the lap is faster than the previous lap, false if it is slower, undefined if it is the first lap
+ * @param lap the lap to display
+ * @param id the lap number
+ */
 export interface LapDisplayProps {
-    faster?: boolean, // true for faster, false for slower, undefined for neither (first lap)
+    faster?: boolean,
     lap: Lap,
     id: number,
 }
 
+/**
+ * Props for the StopWatch component
+ * @param laps the list of laps to display
+ * @param time the current time in milliseconds
+ */
+interface StopWatchProps {
+    laps: Lap[],
+    time: number,
+}
+
+/**
+ * A component that displays a single lap as a list item
+ * Provides a class name of 'faster' or 'slower' if the lap is faster or slower than the previous lap respectively or no class name if it is the first lap
+ * @param props LapDisplayProps
+ * @returns an li element that displays the lap
+ */
 export function LapDisplay (props: LapDisplayProps) {
     return(
         <li
@@ -55,12 +52,16 @@ export function LapDisplay (props: LapDisplayProps) {
     )
 }
 
+/**
+ * A component that displays a list of all the recorded laps
+ * @param props LapListProps
+ * @returns a ul element that contains all the laps as LapDisplays
+ */
 export function LapList(props: LapListProps) {
-
     // This is the lap that is currently being timed
     const previewLap: Lap = {
-        duration: props.time - (props.laps.length > 0 ? props.laps[props.laps.length - 1].startTime : 0),
-        startTime: props.time,
+        duration: props.time - (props.laps.length > 0 ? props.laps[props.laps.length - 1].endTime : 0),
+        endTime: props.time,
     }
 
     return(
@@ -82,6 +83,12 @@ export function LapList(props: LapListProps) {
     )
 }
 
+/**
+ * A component that displays the stopwatch consisting of the LapList and the time counter without any buttons
+ * For buttons see the StopWatchButton component
+ * @param props StopWatchProps
+ * @returns a div element that contains the LapList and the time counter
+ */
 export default function StopWatch(props: StopWatchProps) {
     return(
         <div data-testid="stopwatch" className="stopwatch">
